I have organized the existing file location - Added login/ and register/ subfolders under interface_adapter - Organized use cases into separate Login, Register, PlayerData modules
