+++ core/src/main/java/interface_adapter/login/LoginController.java
@@ -1,4 +1,4 @@
-package interface_adapter;
+package interface_adapter.login;
 import UseCases.Login.LoginInputBoundary;
 
 public class LoginController {
