+++ core/src/main/java/interface_adapter/register/RegisterController.java
@@ -1,4 +1,4 @@
-package interface_adapter;
+package interface_adapter.register;
 
 import UseCases.Register.RegisterInputBoundary;
