+++ core/src/main/java/interface_adapter/register/RegisterViewModel.java
@@ -1,4 +1,4 @@
-package interface_adapter;
+package interface_adapter.register;
 import java.beans.PropertyChangeListener;
 import java.beans.PropertyChangeSupport;
